ya va

+++ Acitvidades/Examen_T3_IvanNR/src/main/java/com/ivan/examen/repository/AulaRepository.java
@@ -2,9 +2,8 @@ package com.ivan.examen.repository;
 
 import com.ivan.examen.model.Aula;
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.List;
 
-public interface AulaRepository extends JpaRepository<Aula, Integer> {
+public interface AulaRepository extends JpaRepository<Aula, Long> {
     List<Aula> findByCapacidadGreaterThan(int capacidad);
 }
