+++ Acitvidades/Examen_T3_IvanNR/src/main/java/com/ivan/examen/repository/CursoRepository.java
@@ -3,6 +3,6 @@ package com.ivan.examen.repository;
 import com.ivan.examen.model.Curso;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CursoRepository extends JpaRepository<Curso, Integer> {
-    Curso findByAulaId(int idAula);
+public interface CursoRepository extends JpaRepository<Curso, Long> {
+    Curso findByAulaId(Long idAula);
 }
