+++ Acitvidades/Examen_T3_IvanNR/src/main/java/com/ivan/examen/repository/ProfesorRepository.java
@@ -3,5 +3,5 @@ package com.ivan.examen.repository;
 import com.ivan.examen.model.Profesor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProfesorRepository extends JpaRepository<Profesor, Integer> {
+public interface ProfesorRepository extends JpaRepository<Profesor, Long> {
 }
