+++ Acitvidades/Examen_T3_IvanNR/src/main/java/com/ivan/examen/service/CursoService.java
@@ -1,7 +1,9 @@
 package com.ivan.examen.service;
 
 import com.ivan.examen.model.Curso;
+import java.util.List;
 
 public interface CursoService {
-    Curso findByAulaId(int idAula);
+
+    Curso findByAulaId(Long idAula);
 }
