+++ Acitvidades/Examen_T3_IvanNR/src/main/java/com/ivan/examen/service/ProfesorCursoService.java
@@ -1,11 +1,9 @@
 package com.ivan.examen.service;
 
 import com.ivan.examen.model.Profesor;
-
 import java.util.List;
 
 public interface ProfesorCursoService {
-    List<Profesor> findProfesoresByCursoId(int idCurso);
-
-    List<String> findCursosByProfesorId(int idProfesor);
+    List<Profesor> findProfesoresByCursoId(Long idCurso);
+    List<String> findCursosByProfesorId(Long idProfesor);
 }
